% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_roxy.R
\name{create_roxy}
\alias{create_roxy}
\alias{make_roxy}
\alias{dir_roxy}
\alias{file_roxy}
\alias{obj_roxy}
\title{Helper functions using \code{{sinew}} to make roxygen2 skeletons for directories,
files, and objects}
\usage{
dir_roxy(path, markdown = TRUE, quiet = FALSE, ...)

file_roxy(path, markdown = TRUE, quiet = FALSE, ...)

obj_roxy(obj, title = NULL, description = NULL, markdown = TRUE, ...)
}
\arguments{
\item{markdown}{boolean to return roxygen2 skeleton with Markdown formatting,
Default: FALSE}

\item{...}{additional parameters passed to \code{makeOxygen}}

\item{obj}{function or name of function}

\item{title, description}{Optional title and description values to use}
}
\description{
These functions are wrappers for \code{\link[sinew:makeOxyFile]{sinew::makeOxyFile()}} and
\code{\link[sinew:makeOxygen]{sinew::makeOxygen()}}.
}
\seealso{
\code{\link[sinew:makeOxyFile]{sinew::makeOxyFile()}}
}
