% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2path-package.R
\docType{package}
\name{roxygen2path-package}
\alias{roxygen2path}
\alias{roxygen2path-package}
\title{roxygen2path: Make Roxygen2 Docs from a File Path or Directory}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://elipousson.github.io/roxygen2path/}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

}
\keyword{internal}
