% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_roxy.R
\name{create_roxy}
\alias{create_roxy}
\alias{make_roxy}
\alias{dir_roxy}
\alias{file_roxy}
\alias{obj_roxy}
\title{Helper functions using \code{{sinew}} to make roxygen2 skeletons for directories,
files, and objects}
\usage{
dir_roxy(path, markdown = TRUE, quiet = FALSE, ...)

file_roxy(path, markdown = TRUE, quiet = FALSE, ...)

obj_roxy(obj, title = NULL, description = NULL, markdown = TRUE, ...)
}
\arguments{
\item{markdown}{boolean to return roxygen2 skeleton with Markdown formatting,
Default: FALSE}

\item{...}{
  Arguments passed on to \code{\link[sinew:makeOxygen]{makeOxygen}}
  \describe{
    \item{\code{add_default}}{boolean to add defaults values to the end of the PARAM
fields, Default: TRUE}
    \item{\code{add_fields}}{character vector to add additional roxygen2 fields,
Default: c("details","examples","seealso","rdname","export")}
    \item{\code{use_dictionary}}{character, path_to_dictionary, Default: NULL}
    \item{\code{use_labels}}{boolean to use label attribute of data frame columns to
fill column description values. Ignored if obj is a function or vector.
Default: FALSE.}
    \item{\code{scope}}{The scope of transformations: \code{"simple"} runs only transformations
that shouldn't substantially change the resulting \code{.Rd} files, \code{"full"} runs
all transformations. In larger packages, run \code{"none"}, double-check and track
the changes, and then run \code{"simple"} and then \code{"full"}.}
  }}

\item{obj}{function or name of function, a data frame, or a vector (strings
are interpreted as function names).}

\item{title, description}{Optional title and description values to use}
}
\description{
These functions are wrappers for \code{\link[sinew:makeOxyFile]{sinew::makeOxyFile()}} and
\code{\link[sinew:makeOxygen]{sinew::makeOxygen()}}.
}
\seealso{
\code{\link[sinew:makeOxyFile]{sinew::makeOxyFile()}}
}
