% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_document.R
\name{create_temp_package}
\alias{create_temp_package}
\title{Create a temporary R package from a script or folder}
\usage{
create_temp_package(
  path,
  pkg = NULL,
  fields = list(),
  rstudio = FALSE,
  check_name = TRUE,
  open = FALSE,
  tmp_dir = tempdir(),
  overwrite = FALSE,
  call = caller_env()
)
}
\arguments{
\item{path}{A character vector of one or more paths.}

\item{pkg}{Package name. Optional.}

\item{fields}{A named list of fields to add to \code{DESCRIPTION}, potentially
overriding default values. See \code{\link[usethis:use_description]{use_description()}} for how you can set
personalized defaults using package options.}

\item{rstudio}{If \code{TRUE}, calls \code{\link[usethis:use_rstudio]{use_rstudio()}} to make the new package or
project into an \href{https://r-pkgs.org/workflow101.html#sec-workflow101-rstudio-projects}{RStudio Project}.
If \code{FALSE} and a non-package project, a sentinel \code{.here} file is placed so
that the directory can be recognized as a project by the
\href{https://here.r-lib.org}{here} or
\href{https://rprojroot.r-lib.org}{rprojroot} packages.}

\item{check_name}{Whether to check if the name is valid for CRAN and throw an
error if not.}

\item{open}{If \code{TRUE}, \link[usethis:proj_activate]{activates} the new project:
\itemize{
\item If using RStudio desktop, the package is opened in a new session.
\item If on RStudio server, the current RStudio project is activated.
\item Otherwise, the working directory and active project is changed.
}}

\item{tmp_dir}{Directory where temporary package is created. Defaults to
\code{\link[=tempdir]{tempdir()}}.}

\item{overwrite}{Default \code{FALSE}. If \code{TRUE}, overwrite the NAMESPACE file and
any R scripts at the same location sharing the same names.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
Invisibly returns path to the package location
}
\description{
If pkg is not provided, the package name is based on the basename of the
project working directory. By default, the package is created in a temporary
directory and deleted when the session restarts.
}
