% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_document.R
\name{path_document}
\alias{path_document}
\alias{path_roxygenise}
\title{Create a temporary package and use \code{\link[=roxygenise]{roxygenise()}} to create and load documentation}
\usage{
path_document(
  path,
  pkg = NULL,
  fields = list(),
  tmp_dir = tempdir(),
  markdown = TRUE,
  overwrite = FALSE,
  roclets = NULL,
  load_code = NULL,
  clean = FALSE,
  quiet = FALSE,
  allow_file = TRUE,
  regexp = "\\\\.[rR]$",
  ...
)

path_roxygenise(
  path,
  pkg = NULL,
  fields = list(),
  tmp_dir = tempdir(),
  markdown = TRUE,
  overwrite = FALSE,
  roclets = NULL,
  load_code = NULL,
  clean = FALSE,
  quiet = FALSE,
  allow_file = TRUE,
  regexp = "\\\\.[rR]$",
  ...
)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created, provided that the parent path exists.}

\item{fields}{A named list of fields to add to \code{DESCRIPTION}, potentially
overriding default values. See \code{\link[usethis:use_description]{use_description()}} for how you can set
personalized defaults using package options.}

\item{regexp}{A regular expression (e.g. \verb{[.]csv$}) passed on to \code{\link[=grep]{grep()}} to filter paths.}

\item{...}{Unused at present.}
}
\description{
Inspired by \code{\link[document:document]{document::document()}}
}
