% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_document.R
\name{pkg_dir_temp}
\alias{pkg_dir_temp}
\title{Create a temporary directory for a R package}
\usage{
pkg_dir_temp(path, tmp_dir = tempdir(), overwrite = FALSE, call = caller_env())
}
\arguments{
\item{path}{A character vector of one or more paths.}

\item{tmp_dir}{The directory the file will be created in.}
}
\description{
Create a temporary directory for a R package
}
